# Make getting style violations thread-safe and parallelize linting