--- a/Source/SwiftLintFramework/Extensions/Array+SwiftLint.swift
+++ b/Source/SwiftLintFramework/Extensions/Array+SwiftLint.swift
@@ -44,43 +44,22 @@ extension Array {
         }
     }
 
-    public func parallelForEach(block: @escaping ((Element) -> Void)) {
-        _ = parallelMap(transform: block)
-    }
-
     func parallelFlatMap<T>(transform: @escaping ((Element) -> [T])) -> [T] {
         return parallelMap(transform: transform).flatMap { $0 }
     }
 
     func parallelMap<T>(transform: @escaping ((Element) -> T)) -> [T] {
-        let count = self.count
-        let maxConcurrentJobs = ProcessInfo.processInfo.activeProcessorCount
-
-        guard count > 1 && maxConcurrentJobs > 1 else {
-            // skip GCD overhead if we'd only run one at a time anyway
-            return map(transform)
-        }
-
-        var result = [(Int, [T])]()
+        var result = [(Int, T)]()
         result.reserveCapacity(count)
-        let group = DispatchGroup()
-        let uuid = NSUUID().uuidString
-        let jobCount = Int(ceil(Double(count) / Double(maxConcurrentJobs)))
 
-        let queueLabelPrefix = "io.realm.SwiftLintFramework.map.\(uuid)"
+        let queueLabelPrefix = "io.realm.SwiftLintFramework.map.\(NSUUID().uuidString)"
         let resultAccumulatorQueue = DispatchQueue(label: "\(queueLabelPrefix).resultAccumulator")
-
-        for jobIndex in stride(from: 0, to: count, by: jobCount) {
-            let queue = DispatchQueue(label: "\(queueLabelPrefix).\(jobIndex / jobCount)")
-            queue.async(group: group) {
-                let jobElements = self[jobIndex..<Swift.min(count, jobIndex + jobCount)]
-                let jobIndexAndResults = (jobIndex, jobElements.map(transform))
-                resultAccumulatorQueue.sync {
-                    result.append(jobIndexAndResults)
-                }
+        DispatchQueue.concurrentPerform(iterations: count) { index in
+            let jobIndexAndResults = (index, transform(self[index]))
+            resultAccumulatorQueue.sync {
+                result.append(jobIndexAndResults)
             }
         }
-        group.wait()
         return result.sorted { $0.0 < $1.0 }.flatMap { $0.1 }
     }
 }