--- a/Source/swiftlint/Commands/LintCommand.swift
+++ b/Source/swiftlint/Commands/LintCommand.swift
@@ -17,29 +17,31 @@ struct LintCommand: CommandProtocol {
     let verb = "lint"
     let function = "Print lint warnings and errors (default command)"
 
-    private static let violationsAccumulatorQueue = DispatchQueue(label: "io.realm.swiftlint.violationsAccumulator")
-
     func run(_ options: LintOptions) -> Result<(), CommandantError<()>> {
         var fileBenchmark = Benchmark(name: "files")
         var ruleBenchmark = Benchmark(name: "rules")
         var violations = [StyleViolation]()
         let configuration = Configuration(options: options)
         let reporter = reporterFrom(options: options, configuration: configuration)
+        let visitorMutationQueue = DispatchQueue(label: "io.realm.swiftlint.lintVisitorMutation")
         return configuration.visitLintableFiles(options) { linter in
             let currentViolations: [StyleViolation]
             if options.benchmark {
                 let start = Date()
                 let (_currentViolations, currentRuleTimes) = linter.styleViolationsAndRuleTimes
                 currentViolations = _currentViolations
-                fileBenchmark.record(file: linter.file, from: start)
-                currentRuleTimes.forEach { ruleBenchmark.record(id: $0, time: $1) }
+                visitorMutationQueue.sync {
+                    fileBenchmark.record(file: linter.file, from: start)
+                    currentRuleTimes.forEach { ruleBenchmark.record(id: $0, time: $1) }
+                    violations += currentViolations
+                }
             } else {
                 currentViolations = linter.styleViolations
+                visitorMutationQueue.sync {
+                    violations += currentViolations
+                }
             }
             linter.file.invalidateCache()
-            LintCommand.violationsAccumulatorQueue.sync {
-                violations += currentViolations
-            }
             reporter.reportViolations(currentViolations, realtimeCondition: true)
         }.flatMap { files in
             if LintCommand.isWarningThresholdBroken(configuration, violations: violations) {