Addresses #1077. Thread sanitizer seems happy with this when running unit tests and linting a few projects I could find, including the Swift standard library.

Generally speeds up linting by ~30% to ~140% on my 2013 MacBook Pro, though the gains should be higher on more recent CPUs.

* **apple/swift:** 218s to 166s (31% faster)
* **realm/realm-cocoa:** 9.6s to 5.6s (71% faster)
* **Carthage/Carthage:** 96s to 53s (80% faster)
* **jpsim/SourceKitten:** 4.2s to 2.1s (100% faster)
* **realm/SwiftLint:** 6.8s to 3.3s (136% faster)

I had initially added a `--parallel` flag to the lint command, but have since removed that option since I haven't run into any issues with it so far, and couldn't come up with a compelling reason to hide it behind an opt-in flag.