@marcelofabri 2016-12-31
I got an exception on one of my runs:

```
Linting 'SchemaTests.swift' (29/41)
[2]    95692 illegal hardware instruction  
/parallel_swiftlint/swiftlint lint
```