@marcelofabri 2016-12-31
And another one:

```
swiftlint(95787,0x7000043ed000) malloc: *** error for object 0x7f8e41915d18: incorrect checksum for freed object - object was probably modified after being freed.
*** set a breakpoint in malloc_error_break to debug
[2]    95787 abort      parallel_swiftlint/swiftlint lint 
```