@marcelofabri 2016-12-31
If it helps, I'm running on a 6-Core 2013 Mac Pro.