@marcelofabri 2016-12-31
Actually this just happens with the `--benchmark` option.