@jpsim 2017-01-01
Yeah, I had neglected to make recording benchmarks thread-safe. Done in 253e1f4.