@norio-nomura 2017-01-01
I have gotten crashes twice on linting:
https://gist.github.com/norio-nomura/e8dc253954ec8299e1e647aad3d99398