@jpsim 2017-01-02
@norio-nomura what project(s) were you linting when you experienced these crashes?