@norio-nomura 2017-01-02
I experienced on Carthage. It happens rarely. But it happened first and second try with this PR.