@jpsim 2017-01-02
Hmmm, I ran this with Thread Sanitizer enabled against Carthage with all submodules recursively cloned 5 times in a row without hitting any issues. Xcode 8.2 with just a bare `swiftlint` command, no extra configuration files or CLI arguments.

The only change I've made was to enable all opt-in rules:

```diff
--- a/Source/SwiftLintFramework/Models/Configuration.swift
+++ b/Source/SwiftLintFramework/Models/Configuration.swift
@@ -91,11 +91,7 @@ public struct Configuration: Equatable {
                 return whitelistRules.contains(type(of: rule).description.identifier)
             }
         } else {
-            rules = configuredRules.filter { rule in
-                let id = type(of: rule).description.identifier
-                if validDisabledRules.contains(id) { return false }
-                return optInRules.contains(id) || !(rule is OptInRule)
-            }
+            rules = configuredRules
         }
```