@jpsim 2017-01-02
I appreciate everyone's reviews on this! Thanks @norio-nomura, @marcelofabri and special guest @krzysztofzablocki 🙏.

I think I'll be merging this despite @norio-nomura's crashes, since neither of us appear to be able to reproduce them now.